simplify render-test
